Do stuff

+++ include/uwin/mem.h
@@ -5,6 +5,6 @@
 extern void* guest_base;
 #define TARGET_ADDRESS_SPACE_SIZE 0x80000000
 
-#define g2h(x) ((void*)(guest_base + (unsigned long)(uint32_t)(x)))
-#define h2g(x) ((uint32_t)((unsigned long)(void*)(x) - (unsigned long)guest_base))
+#define g2h(x) ((void*)((uintptr_t)guest_base + (uintptr_t)(uint32_t)(x)))
+#define h2g(x) ((uint32_t)((uintptr_t)(void*)(x) - (uintptr_t)guest_base))
